fix: improve tests, including their co-location, dependencies and scripts
