+++ cypress.config.ts
@@ -7,14 +7,20 @@ export default defineConfig({
       bundler: "vite",
     },
   },
-
   e2e: {
+    baseUrl: "http://localhost:5173",
+    retries: {
+      runMode: 3,
+      openMode: 2,
+    },
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
     setupNodeEvents(on, config) {
-      return {
-        baseUrl: "http://localhost:5173",
-        // implement node event listeners here
-      };
+      on("task", {
+        log(message) {
+          console.log(message);
+          return null;
+        },
+      });
     },
   },
 });
