+++ cypress/e2e/SignUpForm.spec.cy.tsx
@@ -1,4 +1,4 @@
-import { AuthenticationEndPoints } from "@networking/constants";
+import { AuthenticationEndPoints } from "../../src/networking/constants";
 
 enum WaitTimes {
   Short = 2000,
@@ -16,12 +16,12 @@ const user = {
   username: "testuser",
 };
 
-describe("Registration form", () => {
+describe("Signup form", () => {
   beforeEach(() => {
     cy.visit("/signup");
   });
 
-  it.skip("1. will disable submit button if email or password fails validation", () => {
+  it("1. will disable submit button if email or password fails validation", () => {
     cy.wait(WaitTimes.Long);
 
     //email validation failure
@@ -38,7 +38,7 @@ describe("Registration form", () => {
     cy.get("[data-cy=button]").first().should("not.be.disabled");
   });
 
-  it.skip("2. displays an error message in an alert on API error", () => {
+  it("2. displays an error message in an alert on API error", () => {
     const errorMessage = "User already exists";
 
     cy.wait(WaitTimes.Long);
