+++ cypress/index.ts
@@ -1,0 +1,1 @@
+export { default as withProviders } from "./utils";
