+++ cypress/utils.tsx
@@ -1,11 +1,11 @@
 import React from "react";
-import { mount } from "@cypress/react";
+import { mount } from "@cypress/react18";
 import { StoreProvider } from "easy-peasy";
 import { ThemeProvider } from "styled-components";
-import { NetworkProvider } from "../src/networking/NetworkProvider";
-import { store } from "../src/redux/store";
-import theme from "../src/global/theme";
-import { Router } from "../src/routing/router";
+import { NetworkProvider } from "@networking/NetworkProvider";
+import { store } from "@redux/store";
+import theme from "@global/theme";
+import { Router } from "@routing/router";
 import "bootstrap/dist/css/bootstrap.min.css";
 import "../src/App.css";
