+++ index.html
@@ -5,7 +5,7 @@
     <link
       rel="icon"
       type="image/svg+xml"
-      href="/public/poker-svgrepo-com.svg"
+      href="/poker-svgrepo-com.svg"
     />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Poker App</title>
