+++ package.json
@@ -11,51 +11,49 @@
     "prepare": "husky install",
     "lint": "eslint --cache --fix \"src/**/*.{ts,tsx}\"",
     "preview": "vite preview",
-    "cypress:open": "cypress open",
-    "cypress:run": "cypress run"
+    "cypress:run:e2e": "cypress run --e2e",
+    "cypress:run:component": "cypress run --component"
   },
   "dependencies": {
-    "@cypress/react": "^8.0.2",
-    "@tanstack/react-query": "^5.45.0",
-    "@tanstack/react-query-devtools": "^5.45.0",
-    "axios": "^1.7.2",
-    "bootstrap": "^5.3.3",
-    "cypress": "^13.11.0",
-    "easy-peasy": "^6.0.4",
-    "eslint-config-react-app": "^7.0.1",
-    "react": "^18.2.0",
-    "react-app-rewire-alias": "^1.1.7",
-    "react-app-rewired": "^2.2.1",
-    "react-bootstrap": "^2.0.3",
-    "react-bootstrap-icons": "^1.6.1",
-    "react-click-away-listener": "^2.2.3",
-    "react-dom": "^18.2.0",
-    "react-error-boundary": "^4.0.13",
-    "react-hook-form": "^7.51.5",
-    "react-icons": "^5.2.1",
-    "react-redux": "^9.1.2",
-    "react-router-dom": "^6.0.2",
-    "react-scripts": "^5.0.1",
-    "react-spring": "^9.7.3",
-    "styled-components": "^6.1.11",
-    "web-vitals": "^1.0.1"
+    "@cypress/react18": "^2.0.1",
+    "@tanstack/react-query": "5.45.0",
+    "@tanstack/react-query-devtools": "5.45.0",
+    "axios": "1.7.2",
+    "bootstrap": "5.3.3",
+    "cypress": "13.11.0",
+    "easy-peasy": "6.0.4",
+    "eslint-config-react-app": "7.0.1",
+    "react": "18.2.0",
+    "react-bootstrap": "2.0.3",
+    "react-bootstrap-icons": "1.6.1",
+    "react-click-away-listener": "2.2.3",
+    "react-dom": "18.3.1",
+    "react-error-boundary": "4.0.13",
+    "react-hook-form": "7.51.5",
+    "react-icons": "5.2.1",
+    "react-redux": "9.1.2",
+    "react-router-dom": "6.0.2",
+    "react-scripts": "5.0.1",
+    "react-spring": "9.7.3",
+    "styled-components": "6.1.11",
+    "web-vitals": "1.0.1"
   },
   "devDependencies": {
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.2.22",
-    "@typescript-eslint/eslint-plugin": "^7.2.0",
-    "@typescript-eslint/parser": "^7.2.0",
-    "@vitejs/plugin-react": "^4.2.1",
-    "eslint": "^8.57.0",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.6",
-    "husky": "^9.0.11",
-    "lint-staged": "^15.2.5",
-    "prettier": "^3.3.2",
-    "typescript": "^5.2.2",
-    "vite": "^5.2.0"
+    "@types/jest": "26.0.15",
+    "@types/node": "12.0.0",
+    "@types/react": "18.3.3",
+    "@types/react-dom": "18.2.22",
+    "@typescript-eslint/eslint-plugin": "7.2.0",
+    "@typescript-eslint/parser": "7.2.0",
+    "@vitejs/plugin-react": "4.2.1",
+    "eslint": "8.57.0",
+    "eslint-plugin-react-hooks": "4.6.0",
+    "eslint-plugin-react-refresh": "0.4.6",
+    "husky": "9.0.11",
+    "lint-staged": "15.2.5",
+    "prettier": "3.3.2",
+    "typescript": "5.2.2",
+    "vite": "5.2.0"
   },
   "lint-staged": {
     "*.{ts,tsx}": "eslint --cache --fix",
