+++ src/components/containers/PokerGame/features/hand-evaluators/evaluators/HandEvalutorsContext.ts
@@ -1,4 +1,4 @@
-import type { EvaluatorContext, HandEvaluator } from "./types";
+import type { EvaluatorContext, HandEvaluator } from "../types";
 
 export default class HandEvalutorsContext implements EvaluatorContext {
   constructor(private handEvaluator: HandEvaluator) {}
