+++ src/components/containers/PokerGame/features/hand-evaluators/evaluators/tests/FlushEvaluator.spec.cy.ts
@@ -1,4 +1,4 @@
-import { HandEvalutorsContext, FlushEvaluator } from "../";
+import { HandEvalutorsContext, FlushEvaluator } from "../..";
 import {
   mockFlushSameSuitButInSequence,
   mockHandLegitRoyalFlush,
