+++ src/components/containers/PokerGame/features/hand-evaluators/evaluators/tests/StraighFlushEvaluator.spec.cy.ts
@@ -1,4 +1,4 @@
-import { HandEvalutorsContext, StraighFlushEvaluator } from "../index";
+import { HandEvalutorsContext, StraighFlushEvaluator } from "../../index";
 import {
   mockLegitStraightFlush,
   mockStraightFlushHandNotInSequence,
