+++ src/components/containers/PokerGame/features/hand-evaluators/evaluators/tests/TwoPairEvaluator.spec.cy.ts
@@ -1,4 +1,4 @@
-import { HandEvalutorsContext, TwoPairEvaluator } from "../index";
+import { HandEvalutorsContext, TwoPairEvaluator } from "../../index";
 import {
   mockHandLegitRoyalFlush,
   mockLegitFlush,
