+++ src/components/library/PokerTable/PokerTable.styles.ts
@@ -15,7 +15,6 @@ const StyledPokerTable = styled.div<StyledPokerTableProps>(
       backgroundColor: theme.colors[color],
       borderRadius: "10px",
       boxShadow: "0px 4px 8px rgba(0, 0, 0, 0.2)",
-      marginTop: 32,
     }),
 );
