+++ src/routing/router.tsx
@@ -4,6 +4,7 @@ import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import { PrivateRoutes } from "./PrivateRoutes";
 import { PublicRoutes } from "./PublicRoutes";
 import { useCheckIfUser } from "@features/authentication/useCheckIfUser";
+import { ColorOptions } from "@global/theme";
 
 const LoginPage = lazy(() =>
   import("@pages/index").then((module) => ({
@@ -46,7 +47,9 @@ export const Router: FC<RouterProps> = ({ children }) => {
   return (
     <BrowserRouter>
       {children}
-      <Suspense fallback={<p>Loading...</p>}>
+      <Suspense
+        fallback={<p style={{ color: ColorOptions.Gold }}>Loading...</p>}
+      >
         <Routes>
           <Route element={<PrivateRoutes isAuthenticated={isAuthenticated} />}>
             <Route path={RoutePaths.Game} element={<GamePage />} />
