+++ tsconfig.json
@@ -28,7 +28,18 @@
   },
   "include": [
     "src",
-    "config-overrides.js",
-    "cypress/e2e/SignUpForm.spec.cy.tsx"
+    "cypress/e2e/SignUpForm.spec.cy.tsx",
+    "src/components/library/Alert/Alert.spec.cy.tsx",
+    "src/components/library/Dropdown/Dropdown.spec.cy.tsx",
+    "src/components/library/Button/Button.spec.cy.tsx",
+    "src/components/containers/PokerGame/features/hand-evaluators/TwoPairEvaluator.spec.cy.ts",
+    "src/components/containers/PokerGame/features/hand-evaluators/ThreeOfAKindEvaluator.spec.cy.ts",
+    "src/components/containers/PokerGame/features/hand-evaluators/FlushEvaluator.cy.ts",
+    "src/components/containers/PokerGame/features/hand-evaluators/FourOfAKindEvaluator.spec.cy.ts",
+    "src/components/containers/PokerGame/features/hand-evaluators/FullHouseEvaluator.spec.cy.ts",
+    "src/components/containers/PokerGame/features/hand-evaluators/PairEvaluator.spec.cy.ts",
+    "src/components/containers/PokerGame/features/hand-evaluators/RoyalFlushEvaluator.spec.cy.ts",
+    "src/components/containers/PokerGame/features/hand-evaluators/StraighFlushEvaluator.spec.cy.ts",
+    "src/components/containers/PokerGame/features/hand-evaluators/StraightEvaluator.spec.cy.ts"
   ]
 }
